Isolate some grunt code/config
