+++ grunt/app_config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  tasksDir: [
+    './grunt/tasks',
+    './vendor/open-orchestra/open-orchestra-cms-bundle/GruntTasks',
+    './vendor/open-orchestra/open-orchestra-media-admin-bundle/GruntTasks'
+  ],
+  targetsDir: [
+    './targets/',
+    './vendor/open-orchestra/open-orchestra-cms-bundle/GruntTasks/Targets/',
+    './vendor/open-orchestra/open-orchestra-media-admin-bundle/GruntTasks/Targets/'
+  ]
+};
